Drop support for Deluge 1 and pre-releases of Deluge 2
The only user of this fork is deluge_exporter, which has already dropped
support for Deluge 1. Users that were running pre-releases of Deluge 2
will most likely have upgraded to 2.0.0 or later at this point.

This fixes:

  https://github.com/JohnDoee/deluge-client/issues/40
  https://github.com/tobbez/deluge_exporter/issues/22

There's also a ticked filed against Deluge for adding a proper way to
determine a daemon's protocol version, which (if implemented) would make
it easier to support multiple protocol versions in the future:

https://dev.deluge-torrent.org/ticket/3609